tune knight pst properly
Score of Koivisto Dev vs Koivisto 0.003 speedups: 1216 - 1086 - 1698 [0.516]
...      Koivisto Dev playing White: 640 - 523 - 837  [0.529] 2000
...      Koivisto Dev playing Black: 576 - 563 - 861  [0.503] 2000
...      White vs Black: 1203 - 1099 - 1698  [0.513] 4000
Elo difference: 11.3 +/- 8.2, LOS: 99.7 %, DrawRatio: 42.4 %
4000 of 4000 games finished.

Former-commit-id: 9426317106ca665830b20487fba0c3a33a9f1dab

+++ src_files/Tuning.cpp
@@ -533,7 +533,7 @@ double tuning::optimisePST(Evaluator *evaluator, double K, double learningRate) 
         double      sigPrime        = sigmoidPrime(q_i, K);
         double      lossPrime       = - 2 * (expected - sig);
         
-        double*     features        = evaluator->getFeatures();
+        float*     features        = evaluator->getFeatures();
         double      phase           = evaluator->getPhase();
